requirements.txt
streamlit
google-generativeai
pandas
pillow
qrcode